use graphicalview in all 2d plots
