+++ src/plopp/plotting/slicer.py
@@ -10,13 +10,8 @@ from scipp.typing import VariableLike
 
 from ..core import widget_node
 from ..core.typing import PlottableMulti
-from ..graphics import figure1d, figure2d
-from .common import (
-    input_to_nodes,
-    preprocess,
-    raise_multiple_inputs_for_2d_plot_error,
-    require_interactive_backend,
-)
+from .common import input_to_nodes, preprocess, require_interactive_backend
+from .plot import plot
 
 
 class Slicer:
@@ -123,25 +118,10 @@ class Slicer:
         self.slider_node = widget_node(self.slider)
         self.slice_nodes = [slice_dims(node, self.slider_node) for node in nodes]
 
-        ndims = len(keep)
-        if ndims == 1:
-            make_figure = figure1d
-        elif ndims == 2:
-            if len(self.slice_nodes) > 1:
-                raise_multiple_inputs_for_2d_plot_error(origin='slicer')
-            make_figure = figure2d
-        else:
-            raise ValueError(
-                f'Slicer plot: the number of dims to be kept must be 1 or 2, '
-                f'but {ndims} were requested.'
-            )
-        self.figure = make_figure(
-            *self.slice_nodes,
-            autoscale=autoscale,
-            vmin=vmin,
-            vmax=vmax,
-            **kwargs,
+        self.figure = plot(
+            *self.slice_nodes, autoscale=autoscale, vmin=vmin, vmax=vmax, **kwargs
         )
+        self.figure.bottom_bar.add(self.slider)
 
 
 def slicer(
@@ -188,7 +168,7 @@ def slicer(
         A :class:`Box` which will contain a :class:`Figure` and slider widgets.
     """
     require_interactive_backend('slicer')
-    sl = Slicer(
+    return Slicer(
         obj,
         keep=keep,
         autoscale=autoscale,
@@ -196,7 +176,4 @@ def slicer(
         vmax=vmax,
         coords=coords,
         **kwargs,
-    )
-    from ..widgets import Box
-
-    return Box([sl.figure, sl.slider])
+    ).figure
